:recycle: 合并user和user tag
