+++ src/OfficialAccount/OfficialAccount.php
@@ -22,7 +22,6 @@ class OfficialAccount
     private $media;
     private $material;
     private $user;
-    private $userTag;
     private $accessToken;
     private $onError;
     private $customerService;
@@ -125,15 +124,6 @@ class OfficialAccount
             $this->user = new User($this);
         }
         return $this->user;
-    }
-
-    function userTag():UserTag
-    {
-        if(!isset($this->userTag))
-        {
-            $this->userTag = new UserTag($this);
-        }
-        return $this->userTag;
     }
 
     function getConfig():OfficialAccountConfig
